server/requirements.txt
Django==1.8.6
Markdown==2.6.4
django-bootstrap3==6.2.2
django-filter==0.11.0
djangorestframework==3.3.1
gunicorn==19.3.0
wsgiref==0.1.2
